improved for azure and civo
